+++ src/Schema/Events/Tap.ts
@@ -675,3 +675,25 @@ export interface TappedLink {
   destination_screen_owner_id?: string
   destination_screen_owner_slug?: string
 }
+
+/**
+ * A user taps "Create alert button" on an app
+ *
+ * This schema describes events sent to Segment from [[tappedCreateAlert]]
+ *
+ *  @example
+ *  ```
+ *  {
+ *    action: "tappedCreateAlert",
+ *    context_screen_owner_type: "artist",
+ *    context_screen_owner_id: "58de681f275b2464fcdde097",
+ *    context_screen_owner_slug: "anthony-hunter",
+ *  }
+ * ```
+ */
+export interface TappedCreateAlert {
+  action: ActionType.tappedCreateAlert
+  context_screen_owner_type: ScreenOwnerType
+  context_screen_owner_id?: string
+  context_screen_owner_slug?: string
+}